requirements.txt
python-telegram-bot==20.7
sqlite3
python-dotenv==1.0.0
schedule==1.2.0
stripe==7.9.0
